requirements.txt
fastapi
uvicorn
python-dotenv
spotipy
torch
transformers
Pillow
typing-extensions
sqlalchemy
psycopg2-binary
alembic